trace: Fix parameter types in hw/nvram
An upcoming patch will let the compiler warn us when we are silently
losing precision in traces; update the traces to use types matching
the callers, and in this case, make sure all callers pass the same
width type.

Message-Id: <20170313195547.21466-23-eblake@redhat.com>

+++ hw/nvram/fw_cfg.c
@@ -435,7 +435,7 @@ static void fw_cfg_dma_transfer(FWCfgState *s)
     stl_be_dma(s->dma_as, dma_addr + offsetof(FWCfgDmaAccess, control),
                 dma.control);
 
-    trace_fw_cfg_read(s, 0);
+    trace_fw_cfg_read(s, (uint64_t) 0);
 }
 
 static uint64_t fw_cfg_dma_mem_read(void *opaque, hwaddr addr,
